requirements.txt
accelerate==1.10.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aistudio-sdk==0.3.8
altair==5.5.0
annotated-types==0.7.0
anyascii==0.3.3
anyio==4.9.0
apted==1.0.3
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
av==15.0.0
babel==2.17.0
bce-python-sdk==0.9.46
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
boto3==1.34.102
botocore==1.34.162
cachetools==6.1.0
catalogue==2.0.10
certifi==2025.8.3
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.3
chinese_calendar==1.10.0
click==8.2.1
colorama==0.4.6
colorlog==6.9.0
comm==0.2.3
contourpy==1.3.3
cryptography==45.0.6
cssselect==1.3.0
cssutils==2.11.1
cycler==0.12.1
dataclasses-json==0.6.7
datasets==4.0.0
debugpy==1.8.16
decorator==5.2.1
decord==0.6.0
deepdoctection==0.45.0
defusedxml==0.7.1
dill==0.3.8
Distance==0.1.3
distro==1.9.0
einops==0.8.1
et_xmlfile==2.0.0
executing==2.2.0
faiss-cpu==1.12.0
fastjsonschema==2.21.2
filelock==3.19.1
fonttools==4.59.1
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.3.0
ftfy==6.3.1
future==1.0.0
gitdb==4.0.12
GitPython==3.1.45
GPUtil==1.4.0
greenlet==3.2.4
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.7
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
idna==3.10
imageio==2.37.0
imagesize==1.4.1
importlib_metadata==8.7.0
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
jdeskew==0.3.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
json5==0.12.1
jsonlines==3.1.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.7
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.9
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.74
langchain-openai==0.3.30
langchain-text-splitters==0.3.9
langdetect==1.0.9
langsmith==0.4.14
language_tool_python==2.9.4
lark==1.2.2
lazy-imports==0.3.1
lazy_loader==0.4
lxml==6.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.5
matplotlib-inline==0.1.7
mistune==3.1.4
ml_dtypes==0.5.3
mock==4.0.3
modelscope==1.30.0
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.16
mypy_extensions==1.1.0
narwhals==2.1.2
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
notebook==7.4.5
notebook_shim==0.2.4
numpy==2.2.6
onnx==1.19.0
openai==1.99.9
opencv-contrib-python==4.10.0.84
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
python-docx==1.1.2
xlrd==2.0.1
opt-einsum==3.3.0
orjson==3.11.2
packaging==25.0
paddleocr==3.2.0
paddlepaddle==3.1.1
paddlex==3.2.1
pandas==2.3.1
pandocfilters==1.5.1
parso==0.8.4
pdfminer.six==20250506
pdfplumber==0.11.7
peft==0.17.1
pillow==11.3.0
platformdirs==4.4.0
premailer==3.10.0
prettytable==3.16.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.30.2
psutil==7.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==21.0.0
pyclipper==1.3.0.post6
pycocotools==2.0.8
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.3
pypdf==6.0.0
pypdfium2==4.30.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-doctr==0.10.0
python-dotenv==1.1.1
python-json-logger==3.3.0
pytz==2025.2
pywin32==311
pywinpty==3.0.0
PyYAML==6.0.2
pyzmq==27.0.2
qwen-vl-utils==0.0.11
RapidFuzz==3.14.0
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.27.0
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
s3transfer==0.10.4
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.16.1
Send2Trash==1.8.3
sentencepiece==0.2.1
setuptools==80.9.0
shapely==2.1.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.7
SQLAlchemy==2.0.43
stack-data==0.6.3
streamlit==1.48.1
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.1.0
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2025.6.11
tiktoken==0.11.0
timm==1.0.19
tinycss2==1.4.0
tokenizers==0.21.4
toml==0.10.2
torch==2.8.0
torchaudio==2.8.0+cpu
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers @ git+https://github.com/huggingface/transformers@eb2f9da096fc59dd80ed14d4165719887dfeb792
types-python-dateutil==2.9.0.20250822
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ujson==5.10.0
uri-template==1.3.0
urllib3==2.5.0
validators==0.35.0
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.24.0
